Move SWT specific unit tests to integTest
In order to avoid invalid thread access errors the ant junit task must
be used in order to properly start on the first thread on MacOS.
